move dashboard insights table to separate file
